refactor(database): packages changed for better organization
