+++ src/main/java/me/zypj/revamp/leaderboard/repository/impl/SQLiteBoardRepository.java
@@ -1,7 +1,8 @@
-package me.zypj.revamp.leaderboard.repository;
+package me.zypj.revamp.leaderboard.repository.impl;
 
 import me.zypj.revamp.leaderboard.LeaderboardPlugin;
 import me.zypj.revamp.leaderboard.model.BoardEntry;
+import me.zypj.revamp.leaderboard.repository.BoardRepository;
 
 import javax.sql.DataSource;
 import java.sql.*;
@@ -11,10 +12,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-/**
- * JDBC implementation for MySQL (uses ON DUPLICATE KEY UPDATE).
- */
-public class JdbcBoardRepository implements BoardRepository {
+public class SQLiteBoardRepository implements BoardRepository {
 
     private static final Logger LOGGER = LeaderboardPlugin
             .getPlugin(LeaderboardPlugin.class)
@@ -23,7 +21,7 @@ public class JdbcBoardRepository implements BoardRepository {
     private final DataSource ds;
     private final ExecutorService executor;
 
-    public JdbcBoardRepository(DataSource ds, ExecutorService executor) {
+    public SQLiteBoardRepository(DataSource ds, ExecutorService executor) {
         this.ds = ds;
         this.executor = executor;
     }
@@ -35,11 +33,10 @@ public class JdbcBoardRepository implements BoardRepository {
 
     @Override
     public void initTables(List<String> rawBoards) {
-        String ddlTemplate = "CREATE TABLE IF NOT EXISTS `%s` (" +
-                "player_uuid VARCHAR(36) NOT NULL," +
-                "player_name VARCHAR(16) NOT NULL," +
-                "value DOUBLE NOT NULL," +
-                "PRIMARY KEY (player_uuid)" +
+        String ddlTemplate = "CREATE TABLE IF NOT EXISTS \"%s\" (" +
+                "player_uuid TEXT PRIMARY KEY," +
+                "player_name TEXT NOT NULL," +
+                "value REAL NOT NULL" +
                 ")";
         for (String table : rawBoards) {
             String ddl = String.format(ddlTemplate, table);
@@ -49,8 +46,7 @@ public class JdbcBoardRepository implements BoardRepository {
 
     @Override
     public void save(String table, String uuid, String name, double value) {
-        String sql = "INSERT INTO `" + table + "`(player_uuid, player_name, value) VALUES(?, ?, ?)"
-                + " ON DUPLICATE KEY UPDATE player_name=VALUES(player_name), value=VALUES(value)";
+        String sql = "INSERT OR REPLACE INTO \"" + table + "\"(player_uuid, player_name, value) VALUES(?, ?, ?)";
         executor.submit(() -> {
             try (Connection c = ds.getConnection();
                  PreparedStatement ps = c.prepareStatement(sql)) {
@@ -68,21 +64,20 @@ public class JdbcBoardRepository implements BoardRepository {
     public void batchSave(String table, List<JdbcBoardRepository.BoardBatchEntry> batch) {
         if (batch.isEmpty()) return;
 
-        StringBuilder sb = new StringBuilder("INSERT INTO `")
+        StringBuilder sb = new StringBuilder("INSERT OR REPLACE INTO \"")
                 .append(table)
-                .append("`(player_uuid, player_name, value) VALUES ");
+                .append("\"(player_uuid, player_name, value) VALUES ");
         for (int i = 0; i < batch.size(); i++) {
             sb.append("(?, ?, ?)");
             if (i < batch.size() - 1) sb.append(',');
         }
-        sb.append(" ON DUPLICATE KEY UPDATE player_name=VALUES(player_name), value=VALUES(value)");
         String sql = sb.toString();
 
         executor.submit(() -> {
             try (Connection c = ds.getConnection();
                  PreparedStatement ps = c.prepareStatement(sql)) {
                 int idx = 1;
-                for (BoardBatchEntry e : batch) {
+                for (JdbcBoardRepository.BoardBatchEntry e : batch) {
                     ps.setString(idx++, e.uuid);
                     ps.setString(idx++, e.name);
                     ps.setDouble(idx++, e.value);
@@ -96,14 +91,14 @@ public class JdbcBoardRepository implements BoardRepository {
 
     @Override
     public void truncate(String table) {
-        executor.submit(() -> executeUpdate("TRUNCATE TABLE `" + table + "`"));
+        executor.submit(() -> executeUpdate("DELETE FROM \"" + table + "\""));
     }
 
     @Override
     public List<BoardEntry> loadTop(String table, int limit) {
         List<BoardEntry> list = new ArrayList<>();
         String sql = "SELECT player_uuid, player_name, value "
-                + "FROM `" + table + "` "
+                + "FROM \"" + table + "\" "
                 + "ORDER BY value DESC, player_name ASC"
                 + (limit > 0 ? " LIMIT ?" : "");
 
@@ -133,18 +128,6 @@ public class JdbcBoardRepository implements BoardRepository {
             s.executeUpdate(sql);
         } catch (SQLException ex) {
             LOGGER.log(Level.SEVERE, "Error executing SQL: " + sql, ex);
-        }
-    }
-
-    public static class BoardBatchEntry {
-        public final String uuid;
-        public final String name;
-        public final double value;
-
-        public BoardBatchEntry(String uuid, String name, double value) {
-            this.uuid = uuid;
-            this.name = name;
-            this.value = value;
         }
     }
 }
